Change values in List's FromComfig specs to avoid misleading example
The first and second keys could make someone reading the tests that
conferer is aware of the meaning of those words, but the test only
passed because of how they are ordered lexicographically: `"first" <
"second"`. Replacing first by fourth would still make the test pass, so
I decided to replace them with 0 and 1.